# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AbortRunRequest": {
        "additionalProperties": false,
        "description": "HTTP request body for aborting the active strategy run.",
        "properties": {
          "reason": {
            "default": "user_requested",
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "AbortRunRequest",
        "type": "object"
      },
      "AbortRunResponse": {
        "additionalProperties": false,
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "AbortRunResponse",
        "type": "object"
      },
...
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MemoryDreamResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "target_date": {
            "format": "date",
            "title": "Target Date",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "target_date",
          "status",
          "result",
          "failure_reason",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "MemoryDreamResponse",
        "type": "object"
      },
...
      "RunDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_state": {
            "title": "Current State",
            "type": "string"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "schedule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "summary_render_mode": {
            "default": "markdown",
            "enum": [
              "markdown",
              "html"
            ],
            "title": "Summary Render Mode",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "thinking_count": {
            "default": 0,
            "title": "Thinking Count",
            "type": "integer"
          },
          "tool_calls_count": {
            "default": 0,
            "title": "Tool Calls Count",
            "type": "integer"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          },
          "trace": {
            "$ref": "#/components/schemas/RunTraceResponse"
          },
          "trade_count": {
            "default": 0,
            "title": "Trade Count",
            "type": "integer"
          },
          "trigger_source": {
            "title": "Trigger Source",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "task_id",
          "trigger_source",
          "status",
          "current_state",
          "started_at",
          "trace"
        ],
        "title": "RunDetailResponse",
        "type": "object"
      },
      "RunSummaryResponse": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_state": {
            "title": "Current State",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "schedule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "summary_render_mode": {
            "default": "markdown",
            "enum": [
              "markdown",
              "html"
            ],
            "title": "Summary Render Mode",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "thinking_count": {
            "default": 0,
            "title": "Thinking Count",
            "type": "integer"
          },
          "tool_calls_count": {
            "default": 0,
            "title": "Tool Calls Count",
            "type": "integer"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          },
          "trade_count": {
            "default": 0,
            "title": "Trade Count",
            "type": "integer"
          },
          "trigger_source": {
            "title": "Trigger Source",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "task_id",
          "trigger_source",
          "status",
          "current_state",
          "started_at"
        ],
        "title": "RunSummaryResponse",
        "type": "object"
      },
      "RunTraceResponse": {
        "additionalProperties": false,
        "properties": {
          "current_stage_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage Id"
          },
          "event_seq": {
            "minimum": 0.0,
            "title": "Event Seq",
            "type": "integer"
          },
          "schema_version": {
            "const": 3,
            "title": "Schema Version",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/TraceStageResponse"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "event_seq",
          "current_stage_id",
          "stages"
        ],
        "title": "RunTraceResponse",
        "type": "object"
      },
...
      "StartRunRequest": {
        "additionalProperties": false,
        "description": "HTTP request body for starting a strategy run.",
        "properties": {
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_version": {
            "default": "m1-bootstrap",
            "title": "Prompt Version",
            "type": "string"
          },
          "risk_rules_version": {
            "default": "m1-st-only",
            "title": "Risk Rules Version",
            "type": "string"
          },
          "schedule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Id"
          },
          "trigger_source": {
            "$ref": "#/components/schemas/TriggerSource",
            "default": "manual"
          }
        },
        "title": "StartRunRequest",
        "type": "object"
      },
...
      "TraceStageResponse": {
        "additionalProperties": false,
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "key": {
            "enum": [
              "run",
              "summary"
            ],
            "title": "Key",
            "type": "string"
          },
          "stage_id": {
            "title": "Stage Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "completed",
              "degraded",
              "failed",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TraceStepResponse"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "stage_id",
          "key",
          "status",
          "started_at",
          "ended_at",
          "steps"
        ],
        "title": "TraceStageResponse",
        "type": "object"
      },
...
      "TriggerSource": {
        "description": "Supported strategy run trigger sources.",
        "enum": [
          "manual",
          "scheduled"
        ],
        "title": "TriggerSource",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "AniuSessionCookie": {
        "in": "cookie",
        "name": "aniu_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "基于股票交易的可进化智能体系统",
    "title": "Aniu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/aniu/memory-dreams/run": {
      "post": {
        "operationId": "run_memory_dream_api_aniu_memory_dreams_run_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDreamResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Run Memory Dream",
        "tags": [
          "MemoryDreams"
        ]
      }
    },
...
    "/api/aniu/runs": {
      "get": {
        "description": "List strategy runs.",
        "operationId": "list_runs_api_aniu_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "started_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Started Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryResponse"
                  },
                  "title": "Response List Runs Api Aniu Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/aniu/runs/start": {
      "post": {
        "description": "Create one strategy run and queue it on the process-local worker.",
        "operationId": "start_run_api_aniu_runs_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StartRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Start Run",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/aniu/runs/{run_id}": {
      "delete": {
        "description": "Delete one completed or failed run snapshot.",
        "operationId": "delete_run_api_aniu_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Delete Run",
        "tags": [
          "Runs"
        ]
      },
      "get": {
        "description": "Return one run detail record.",
        "operationId": "get_run_detail_api_aniu_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Run Detail",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/aniu/runs/{run_id}/abort": {
      "post": {
        "description": "Request cooperative cancellation for the active run.",
        "operationId": "abort_run_api_aniu_runs__run_id__abort_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AbortRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbortRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Abort Run",
        "tags": [
          "Runs"
        ]
      }
    },
...
    "/api/aniu/sse/run/{run_id}": {
      "get": {
        "operationId": "stream_run_snapshot_api_aniu_sse_run__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent run checkpoint and trace events"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Stream Run Snapshot",
        "tags": [
          "SSE"
        ]
      }
    },
...
  }
}